+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/PackageStoreFactory.java
@@ -30,8 +30,10 @@ import com.vmware.pscoe.iac.artifact.aria.logs.rest.v2.RestClientVrliV2;
 import com.vmware.pscoe.iac.artifact.aria.logs.store.v1.VrliPackageStoreV1;
 import com.vmware.pscoe.iac.artifact.aria.logs.store.v2.VrliPackageStoreV2;
 import com.vmware.pscoe.iac.artifact.aria.operations.configuration.ConfigurationVrops;
+import com.vmware.pscoe.iac.artifact.aria.automation.configuration.ConfigurationAbx;
 import com.vmware.pscoe.iac.artifact.aria.operations.rest.RestClientVrops;
 import com.vmware.pscoe.iac.artifact.aria.operations.store.VropsPackageStore;
+import com.vmware.pscoe.iac.artifact.aria.automation.store.AbxPackageStore;
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.configuration.ConfigurationVro;
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.configuration.ConfigurationVroNg;
 import com.vmware.pscoe.iac.artifact.aria.orchestrator.rest.RestClientVro;
