+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/automation/configuration/ConfigurationAbx.java
@@ -12,9 +12,9 @@
  * This product may include a number of subcomponents with separate copyright notices and license terms. Your use of these subcomponents is subject to the terms and conditions of the subcomponent's license, as noted in the LICENSE file.
  * #L%
  */
-package com.vmware.pscoe.iac.artifact.configuration;
+package com.vmware.pscoe.iac.artifact.aria.automation.configuration;
 
-import com.vmware.pscoe.iac.artifact.aria.automation.configuration.ConfigurationVraNg;
+import com.vmware.pscoe.iac.artifact.configuration.ConfigurationException;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
