+++ common/artifact-manager/src/main/java/com/vmware/pscoe/iac/artifact/aria/automation/rest/RestClientVraNg.java
@@ -23,9 +23,6 @@ import java.util.stream.Collectors;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-import com.vmware.pscoe.iac.artifact.model.abx.AbxAction;
-import com.vmware.pscoe.iac.artifact.model.abx.AbxActionVersion;
-import com.vmware.pscoe.iac.artifact.model.abx.AbxConstant;
 
 import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgApprovalPolicy;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgBlueprint;
@@ -49,6 +46,9 @@ import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgResourceQuotaPo
 import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgSecret;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgSubscription;
 import com.vmware.pscoe.iac.artifact.aria.automation.models.VraNgWorkflowContentSource;
+import com.vmware.pscoe.iac.artifact.aria.automation.models.abx.AbxAction;
+import com.vmware.pscoe.iac.artifact.aria.automation.models.abx.AbxActionVersion;
+import com.vmware.pscoe.iac.artifact.aria.automation.models.abx.AbxConstant;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
